+++ src/pages/error/error.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="error-page-not-found">
     <nav-bar />
-    <div class="pt-10 px-20 <sm:(px-0 pt-0)">
+    <div class="px-20 <sm:(px-0 pt-0)">
       <div class="bg-gray shadow overflow-hidden">
         <div class="px-4 py-5 text-sm font-medium whitespace-nowrap text-white">
           Page not found